Segregation of massive EmeraldTextField protocol
