+++ EmeraldIOS/Protocols/Requirable.swift
@@ -1,0 +1,15 @@
+//
+//  Requirable.swift
+//  EmeraldIOS
+//
+//  Created by Sergio Giraldo on 2/05/20.
+//  Copyright © 2020 Condor Labs. All rights reserved.
+//
+
+import Foundation
+
+public protocol Requirable {
+    func set(isRequired: Bool)
+    func getIsRequired() -> Bool
+    func isValid() -> Result<Bool, Error>
+}
